+++ packages/mui/src/AddButton/AddButton.tsx
@@ -1,6 +1,14 @@
 import AddIcon from '@mui/icons-material/Add';
-import IconButton from '@mui/material/IconButton';
-import { FormContextType, IconButtonProps, RJSFSchema, StrictRJSFSchema, TranslatableString } from '@rjsf/utils';
+import IconButton, { IconButtonProps as MuiIconButtonProps } from '@mui/material/IconButton';
+import {
+  FormContextType,
+  getUiOptions,
+  IconButtonProps,
+  RJSFSchema,
+  StrictRJSFSchema,
+  TranslatableString,
+} from '@rjsf/utils';
+import { getMuiProps } from '../util';
 
 /** The `AddButton` renders a button that represent the `Add` action on a form
  */
@@ -10,8 +18,20 @@ export default function AddButton<T = any, S extends StrictRJSFSchema = RJSFSche
   ...props
 }: IconButtonProps<T, S, F>) {
   const { translateString } = registry;
+  const uiOptions = getUiOptions<T, S, F>(uiSchema);
+  const muiProps = getMuiProps<T, S, F>({
+    uiSchema,
+    formContext: registry.formContext,
+    options: uiOptions,
+  });
+
   return (
-    <IconButton title={translateString(TranslatableString.AddItemButton)} {...props} color='primary'>
+    <IconButton
+      title={translateString(TranslatableString.AddItemButton)}
+      {...props}
+      color='primary'
+      {...(muiProps as MuiIconButtonProps)}
+    >
       <AddIcon />
     </IconButton>
   );