+++ packages/mui/src/DescriptionField/DescriptionField.tsx
@@ -1,6 +1,7 @@
-import Typography from '@mui/material/Typography';
-import { DescriptionFieldProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
+import Typography, { TypographyProps } from '@mui/material/Typography';
+import { DescriptionFieldProps, FormContextType, RJSFSchema, StrictRJSFSchema, getUiOptions } from '@rjsf/utils';
 import { RichDescription } from '@rjsf/core';
+import { getMuiProps } from '../util';
 
 /** The `DescriptionField` is the template to use to render the description of a field
  *
@@ -12,9 +13,24 @@ export default function DescriptionField<
   F extends FormContextType = any,
 >(props: DescriptionFieldProps<T, S, F>) {
   const { id, description, registry, uiSchema } = props;
+
+  const uiOptions = getUiOptions<T, S, F>(uiSchema);
+  const muiProps = getMuiProps<T, S, F>({
+    uiSchema,
+    formContext: registry.formContext,
+    options: uiOptions,
+  });
+  const { slotProps: muiSlotProps, ...otherMuiProps } = muiProps;
+
   if (description) {
     return (
-      <Typography id={id} variant='subtitle2' style={{ marginTop: '5px' }}>
+      <Typography
+        id={id}
+        variant='subtitle2'
+        style={{ marginTop: '5px' }}
+        {...(otherMuiProps as TypographyProps)}
+        {...(muiSlotProps?.typography as TypographyProps)}
+      >
         <RichDescription description={description} registry={registry} uiSchema={uiSchema} />
       </Typography>
     );