+++ packages/mui/src/FieldErrorTemplate/FieldErrorTemplate.tsx
@@ -1,7 +1,8 @@
-import ListItem from '@mui/material/ListItem';
-import FormHelperText from '@mui/material/FormHelperText';
-import List from '@mui/material/List';
-import { errorId, FieldErrorProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
+import ListItem, { ListItemProps } from '@mui/material/ListItem';
+import FormHelperText, { FormHelperTextProps } from '@mui/material/FormHelperText';
+import List, { ListProps } from '@mui/material/List';
+import { errorId, FieldErrorProps, FormContextType, RJSFSchema, StrictRJSFSchema, getUiOptions } from '@rjsf/utils';
+import { getMuiProps } from '../util';
 
 /** The `FieldErrorTemplate` component renders the errors local to the particular field
  *
@@ -12,18 +13,36 @@ export default function FieldErrorTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldErrorProps<T, S, F>) {
-  const { errors = [], fieldPathId } = props;
+  const { errors = [], fieldPathId, uiSchema, registry } = props;
   if (errors.length === 0) {
     return null;
   }
   const id = errorId(fieldPathId);
 
+  const uiOptions = getUiOptions<T, S, F>(uiSchema);
+  const muiProps = getMuiProps<T, S, F>({
+    uiSchema,
+    formContext: registry.formContext,
+    options: uiOptions,
+  });
+  const { slotProps: muiSlotProps, ...otherMuiProps } = muiProps;
+
   return (
-    <List id={id} dense={true} disablePadding={true}>
+    <List
+      id={id}
+      dense={true}
+      disablePadding={true}
+      {...(otherMuiProps as ListProps)}
+      {...(muiSlotProps?.list as ListProps)}
+    >
       {errors.map((error, i: number) => {
         return (
-          <ListItem key={i} disableGutters={true}>
-            <FormHelperText component='div' id={`${id}-${i}`}>
+          <ListItem key={i} disableGutters={true} {...(muiSlotProps?.listItem as ListItemProps)}>
+            <FormHelperText
+              component='div'
+              id={`${id}-${i}`}
+              {...(muiSlotProps?.formHelperText as FormHelperTextProps)}
+            >
               {error}
             </FormHelperText>
           </ListItem>